Refactor gfx resources - remove generic resource system.
Generic resource system was pointless and was making it more awkward to quickly iterate. Since the project is meant to be a testbed I have ditched it in favour of bespoke resource management for the small amount of resources that will be used. This will make it easier to handle bindless etc.